# Conflicts:
#	Assets/Script/Enemys/EnemyBase.cs

+++ Assets/Script/Enemys/EnemyBase.cs
@@ -141,7 +141,7 @@ public abstract class EnemyBase : MonoBehaviour, IPoolable, IDamageable
         }
     }
 
-    public void TakeDamage(float amount, Vector3 hitPoint, Vector3 hitNormal)
+    public void TakeDamage(int amount, Vector3 hitPoint, Vector3 hitNormal)
     {
         if (isDead) return;
 
@@ -186,5 +186,4 @@ public abstract class EnemyBase : MonoBehaviour, IPoolable, IDamageable
             }
         }
     }
-
 }
